requirements.txt
azure-ai-inference>=1.0.0
azure-identity>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
prometheus-client>=0.19.0